% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTTAnalysis.R
\name{Calculate_Skill_Sequence_for_Dataframes}
\alias{Calculate_Skill_Sequence_for_Dataframes}
\title{Calculate_Skill_Sequence_for_Dataframes}
\usage{
Calculate_Skill_Sequence_for_Dataframes(
  data,
  format = "Long",
  number_of_rounds = 3,
  rounds_of_equal_length = TRUE,
  design = c(50, 300, 50),
  automatically_select_span = FALSE,
  span_random = 0.2,
  span_training = 0.14,
  size_of_trial_window_for_skill_same = TRUE,
  size_of_trial_window_for_skill_percentage = 0.28,
  size_of_trial_window_for_skill_percentage_SEQ = 0.1,
  size_of_trial_window_for_skill_percentage_RND = 0.1
)
}
\arguments{
\item{data}{Dataframe including participant data, and their ID}

\item{format}{How are the data shaped? "Long" or "Wide"}

\item{number_of_rounds}{How many rounds did the participants perform? Enter an integer}

\item{rounds_of_equal_length}{Did the participant perform the same amount of trials in every block? TRUE/FALSE}

\item{design}{What was the design of your task? Enter the number of trials of pre-Random, Sequential and post-Random, in this order. Enter a vector (using 'c()') of integer values}

\item{automatically_select_span}{Do you want to manually select the span for the loess function or do you want to automatically select it via generalized cross-validation? TRUE/FALSE}

\item{span_random}{How wide do you want the window of the (pre & post)random phase to be? Numeric between 0 and 1}

\item{span_training}{How wide do you want the window of the sequential phase to be? Numeric between 0 and 1}

\item{size_of_trial_window_for_skill_same}{Do you want to use the same window for calculating Skill for both random and sequential phase? TRUE/FALSE}

\item{size_of_trial_window_for_skill_percentage}{If you want to use the same window, select a numeric value between 0 and 1}

\item{size_of_trial_window_for_skill_percentage_SEQ}{If you DO NOT want to use the same window, select a numeric value between 0 and 1 for the sequential part}

\item{size_of_trial_window_for_skill_percentage_RND}{If you DO NOT want to use the same window, select a numeric value between 0 and 1 for the random part}
}
\value{
A file in wide format that calculate the acquired skill for each round
}
\description{
This function is a wrapper for 'Calculate_Skill_Sequence' to handle dataframes as input data
Going over the 'data' file.
We expect 'data' to have one column for each subject.
We handle also the option where the format is wide
The first row has subject's ID ("ID")
All the other rows are reaction_times
Need to indicate the number of rounds
Each round is split into three sections: Random_Pre, Training and Random_Post
If all rounds have the same number of trials for each section (rounds_of_equal_length = TRUE') (e.g., Random_Pre always 50 trials, Training always 300 trials and Random_Post always 50 trials), only enter once 'design' what the value is for each section.
If rounds have different number of trials for each section (rounds_of_equal_length = FALSE) (e.g., Training at round-1 is 200 trials, then at round-2 300 trials and then 400 trials), you need to specify the entirety of the design (e.g., c(50, 200, 50, 50, 300, 50, 50, 400, 50)).
Random_Pre, Training and Random_Post are free to change.
"automatically_select_span" runs the loess.as	function from the fANCOVA package, which fits a local polynomial regression with automatic smoothing parameter selection via generalized cross-validation
The default value for 'span_random' is 20\% of the total number of trials
The default value for 'span_training' is 14\% of the total number of trials
"size_of_trial_window_for_skill_same" do you want the size of the window to calculate the skill to be the same for the SEQ and the RND?
If TRUE, it will use: "size_of_trial_window_for_skill_percentage", which is the number of trials to be used for calculating the skill. We generally use the last 50 trials of a 180 sequential section (~28\%).
If "size_of_trial_window_for_skill_same" is FALSE, you will need to specify "size_of_trial_window_for_skill_percentage_SEQ" and size_of_trial_window_for_skill_percentage_RND"
}
\examples{
# Basic usage
random_matrix <- matrix(stats::rnorm(10 * 1200, mean = 1, sd = 0.1), nrow = 10, ncol = 1200)
random_df_w <- data.frame(ID = LETTERS[1:10], random_matrix)
Calculate_Skill_Sequence_for_Dataframes(data = random_df_w, format= "Wide" ,  rounds_of_equal_length = FALSE, design = c(50, 200, 50,  50, 300, 50,  50, 400, 50), span_random = 0.20,  span_training = 0.14, automatically_select_span = TRUE, size_of_trial_window_for_skill_same = TRUE, size_of_trial_window_for_skill_percentage = 0.15 , size_of_trial_window_for_skill_percentage_SEQ = 0.28, size_of_trial_window_for_skill_percentage_RND = 0.50)
random_df_l <- as.data.frame(t(random_df_w))
Calculate_Skill_Sequence_for_Dataframes(data = random_df_l, format= "Long" ,  number_of_rounds = 3, rounds_of_equal_length = TRUE, design = c(50, 300, 50), span_random = 0.10,  span_training = 0.4, automatically_select_span = FALSE, size_of_trial_window_for_skill_same = FALSE, size_of_trial_window_for_skill_percentage = 0.15 , size_of_trial_window_for_skill_percentage_SEQ = 0.8, size_of_trial_window_for_skill_percentage_RND = 0.50)
}
