% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTTAnalysis.R
\name{percentage_of_filtered_values_for_Dataframe}
\alias{percentage_of_filtered_values_for_Dataframe}
\title{percentage_of_filtered_values_for_Dataframe}
\usage{
percentage_of_filtered_values_for_Dataframe(
  data,
  format = "Long",
  number_of_rounds = 3,
  rounds_of_equal_length = FALSE,
  design = c(50, 180, 50, 50, 300, 50, 50, 180, 50, 50, 180, 50)
)
}
\arguments{
\item{data}{Dataframe including participant data, and their ID}

\item{format}{How are the data shaped? "Long" or "Wide"}

\item{number_of_rounds}{How many rounds did the participants perform? Enter an integer}

\item{rounds_of_equal_length}{Did the participant perform the same amount of trials in every block? TRUE/FALSE}

\item{design}{What was the design of your task? Enter the number of trials of pre-Random, Sequential and post-Random, in this order. Enter a vector (using 'c()') of integer values}
}
\value{
A file in wide format that gives the percentage of NA values for each section of the design
}
\description{
This function is a wrapper for 'SRTT_analysis_for_Dataframes' to handle dataframes as input data
Going over the 'data' file.
We expect 'data' to have one column for each subject, with ID on top.
We handle also the option where the format is wide
The first row has subject's ID ("ID")
All the other rows are reaction_times
Need to indicate the number of rounds
Each round is split into three sections: Random_Pre, Training and Random_Post
If all rounds have the same number of trials for each section (rounds_of_equal_length = TRUE) (e.g., Random_Pre always 50 trials, Training always 300 trials and Random_Post always 50 trials), only enter once the 'design' values
If rounds have different number of trials for each section (rounds_of_equal_length = FALSE) (e.g., Training at round-1 is 200 trials, then at round-2 300 trials and then 400 trials), you need to specify the entirety of the design (e.g., c(50, 200, 50, 50, 300, 50, 50, 400, 50)).
}
\examples{
# Basic usage
random_matrix <- matrix(stats::rnorm(5 * 1200, mean = 1, sd = 0.1), nrow = 5, ncol = 1200)
for (i in 1:5) { num_nas <- sample(50:300, 1)
random_matrix[i, sample(1200, num_nas)] <- NA }
random_df_w <- data.frame(ID = LETTERS[1:5], random_matrix)
percentage_of_filtered_values_for_Dataframe(random_df_w, format = "Wide", number_of_rounds = 3, rounds_of_equal_length = FALSE, design = c(50 , 180, 50, 50, 300, 50, 50 , 180, 50, 50 , 180, 50))
random_df_l <- as.data.frame(t(random_df_w))
percentage_of_filtered_values_for_Dataframe(data = random_df_l,format = "Long", number_of_rounds = 3, rounds_of_equal_length = FALSE, design = c(50 , 180, 50, 50, 300, 50, 50 , 180, 50, 50 , 180, 50))
}
