% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTTAnalysis.R
\name{plot_data}
\alias{plot_data}
\title{plot_data}
\usage{
plot_data(
  data,
  format = "Wide",
  limit_y_axis_min_max = FALSE,
  limit_y_axis_lower = 0,
  limit_y_axis_upper = 1,
  design = c(50, 200, 50, 50, 100, 50),
  function_to_fit = "loess",
  span_loess_fitting = 0.25,
  standard_error = FALSE,
  color_dots = "black",
  color_line = "grey",
  color_intercept = "grey"
)
}
\arguments{
\item{data}{Single participant data, including their ID}

\item{format}{How are the data shaped? "Long" or "Wide"}

\item{limit_y_axis_min_max}{Do you want to limit the range of the y-axis? TRUE/FALSE}

\item{limit_y_axis_lower}{If you want to set a limit, enter a numeric value for the upper boundary}

\item{limit_y_axis_upper}{If you want to set a limit, enter a numeric value for the lower boundary}

\item{design}{What was the design of your task? Enter the number of trials of pre-Random, Sequential and post-Random, in this order. Enter a vector (using 'c()') of integer values}

\item{function_to_fit}{What function do you want to fit on the data? "loess" or "lm}

\item{span_loess_fitting}{What span do you want to use when fitting a loess function? Enter a value between 0 and 1}

\item{standard_error}{Do you want to show the standard error as a shaded area? TRUE/FALSE}

\item{color_dots}{What color do you want to use for the dots? (e.g., "grey")}

\item{color_line}{What color do you want to use for the fitted line? (e.g., "black")}

\item{color_intercept}{What color do you want to use for the vertical intercept lines? (e.g., "black")}
}
\value{
A figure of all response times
}
\description{
this is a function to visualize response times
each dot is a response
X-axis = time
Y-axis = response time
can fit either a linear or a loess function the data
}
\examples{
# Basic usage
random_list_w <- cbind(999, (as.data.frame(t(stats::rnorm(500, mean = 1, sd = 0.1)))))
random_list_l <- rbind(999 , (as.data.frame(stats::rnorm(500, mean = 1, sd = 0.1))))
plot <- plot_data(random_list_w,format = "Wide" , limit_y_axis_min_max = TRUE,  limit_y_axis_lower = 0.5,  limit_y_axis_upper = 1.5,  design = c(50, 200, 50, 50, 100, 50), function_to_fit = "loess", span_loess_fitting = 0.1,   standard_error = FALSE, color_dots = "black",   color_line = "grey",  color_intercept = "grey")
plot+ggplot2::geom_hline(yintercept =1, col="red")
plot <- plot_data(random_list_l, format = "Long" , limit_y_axis_min_max = TRUE,limit_y_axis_lower = 0.5,  limit_y_axis_upper = 1.5,  design = c(50, 200, 50, 50, 100, 50), function_to_fit = "lm",   span_loess_fitting = 0.1, standard_error = TRUE,  color_dots = "black",  color_line = "grey",  color_intercept = "grey")
plot+ggplot2::geom_hline(yintercept =1, col="red")
}
