% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTTAnalysis.R
\name{plot_data_for_Dataframes}
\alias{plot_data_for_Dataframes}
\title{plot_data_for_Dataframes}
\usage{
plot_data_for_Dataframes(
  data,
  format = "Wide",
  limit_y_axis_min_max = FALSE,
  limit_y_axis_lower = 0,
  limit_y_axis_upper = 1,
  design = c(50, 200, 50, 50, 100, 50),
  function_to_fit = "loess",
  span_loess_fitting = 0.25,
  standard_error = FALSE,
  color_dots = "black",
  color_line = "grey",
  color_intercept = "grey"
)
}
\arguments{
\item{data}{Dataframe including participant data, and their ID}

\item{format}{How are the data shaped? "Long" or "Wide"}

\item{limit_y_axis_min_max}{Do you want to limit the range of the y-axis? TRUE/FALSE}

\item{limit_y_axis_lower}{If you want to set a limit, enter a numeric value for the upper boundary}

\item{limit_y_axis_upper}{If you want to set a limit, enter a numeric value for the lower boundary}

\item{design}{What was the design of your task? Enter the number of trials of pre-Random, Sequential and post-Random, in this order. Enter a vector (using 'c()') of integer values}

\item{function_to_fit}{What function do you want to fit on the data? "loess" or "lm}

\item{span_loess_fitting}{What span do you want to use when fitting a loess function? Enter a value between 0 and 1}

\item{standard_error}{Do you want to show the standard error as a shaded area? TRUE/FALSE}

\item{color_dots}{What color do you want to use for the dots? (e.g., "grey")}

\item{color_line}{What color do you want to use for the fitted line? (e.g., "black")}

\item{color_intercept}{What color do you want to use for the vertical intercept lines? (e.g., "black")}
}
\value{
A figure of all response times
}
\description{
This function is a wrapper for 'plot_data' to handle dataframes as input data
this is a function to visualize response times
each dot is a response
X-axis = time
Y-axis = response time
can fit either a linear or a loess function the data
}
\examples{
# Basic usage
#' random_df_w <- data.frame( ID = LETTERS[1:10], matrix(stats::rnorm(10 * 500, mean = 1, sd = 0.1), nrow = 10, ncol = 500))
plot_data_for_Dataframes(random_df_w, format = "Wide" , # Long limit_y_axis_min_max = FALSE, limit_y_axis_lower = 0.8, limit_y_axis_upper = 1, design = c(50, 200, 50, 50, 100, 50), function_to_fit = "loess", # lm span_loess_fitting = 0.25, standard_error = FALSE, color_dots = "black", color_line = "grey", color_intercept = "grey")
random_df_l <- as.data.frame(t(random_df_w))
plot_data_for_Dataframes(random_df_l,format = "Long" ,  limit_y_axis_min_max = FALSE, limit_y_axis_lower = 0.8,   limit_y_axis_upper = 1,  design = c(50, 100, 50, 50, 100, 50,  25, 50, 25), function_to_fit = "lm",  span_loess_fitting = 0.25,   standard_error = FALSE,  color_dots = "black", color_line = "grey",  color_intercept = "grey")
}
