% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTTAnalysis.R
\name{plot_skills}
\alias{plot_skills}
\title{plot_skills}
\usage{
plot_skills(data, format = "Wide", col_line = "grey", col_dots = "grey")
}
\arguments{
\item{data}{Single participant data, including their ID}

\item{format}{How are the data shaped? "Long" or "Wide"}

\item{col_line}{What color do you want to use for the line? (e.g., "black")}

\item{col_dots}{What color do you want to use for the dots? (e.g., "grey")}
}
\value{
A figure of all Skill
}
\description{
this is a function to visualize Skill
each dot is a response
X-axis = type of skill
Y-axis = response time
}
\examples{
# Basic usage
plot_skills(data, format = "Wide", col_line = "grey",col_dots = "grey")
}
