% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTTAnalysis.R
\name{plot_skills}
\alias{plot_skills}
\title{plot_skills}
\usage{
plot_skills(data, format = "Wide", col_line = "grey", col_dots = "grey")
}
\arguments{
\item{data}{Single participant data, including their ID}

\item{format}{How are the data shaped? "Long" or "Wide"}

\item{col_line}{What color do you want to use for the line? (e.g., "black")}

\item{col_dots}{What color do you want to use for the dots? (e.g., "grey")}
}
\value{
A figure of all Skill
}
\description{
this is a function to visualize Skill
each dot is a response
X-axis = type of skill
Y-axis = response time
}
\examples{
# Basic usage
data_wide <- data.frame(id = 999, t(rnorm(12, mean = 1, sd = 0.1)))
names(data_wide) <- c("ID", "preRND1", "Train1", "postRND1", "Skill1", "preRND2", "Train2", "postRND2", "Skill2", "preRND3", "Train3", "postRND3", "Skill3")
plot_skills(data_wide, format = "Wide", col_line = "grey",col_dots = "grey")
}
