% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTTAnalysis.R
\name{plot_skills_for_Dataframes}
\alias{plot_skills_for_Dataframes}
\title{plot_skills_for_Dataframes}
\usage{
plot_skills_for_Dataframes(
  data,
  format = "Wide",
  col_line = "grey",
  col_dots = "grey",
  error_measure = "se"
)
}
\arguments{
\item{data}{Dataframe including participant data, and their ID}

\item{format}{How are the data shaped? "Long" or "Wide"}

\item{col_line}{What color do you want to use for the line? (e.g., "black")}

\item{col_dots}{What color do you want to use for the dots? (e.g., "grey")}

\item{error_measure}{How do you want to calculate the group-wise error measure? "sd" for standard deviation or "se" for standard error of the mean}
}
\value{
A figure of all Skill
}
\description{
This function is a wrapper for 'plot_skills' to handle dataframes as input data
this is a function to visualize Skill
each dot is a response
X-axis = type of skill
Y-axis = response time
}
\examples{
# Basic usage
random_matrix <- matrix(stats::rnorm(5 * 12, mean = 1, sd = 0.1), nrow = 5, ncol = 12)
random_df_w <- data.frame(ID = LETTERS[1:5], random_matrix)
names(random_df_w) <- c("ID", "preRND1", "Train1", "postRND1", "Skill1", "preRND2", "Train2", "postRND2", "Skill2", "preRND3", "Train3", "postRND3", "Skill3")
plot_skills_for_Dataframes(random_df_w,  format = "Wide",   col_line = "blue", col_dots = "grey")
random_df_l <- as.data.frame(t(random_df_w))
plot_skills_for_Dataframes(random_df_l,  format = "Long",  col_line = "blue",   col_dots = "grey")
}
