% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRTTAnalysis.R
\name{SRTT_analysis_for_Dataframes}
\alias{SRTT_analysis_for_Dataframes}
\title{SRTT_analysis_for_Dataframes}
\usage{
SRTT_analysis_for_Dataframes(
  data,
  format = "Long",
  unit = "Seconds",
  type_of_filtering = "Rolling_Standard_Deviation",
  lower_boundary_fixed_filtering_milliseconds = 200,
  upper_boundary_fixed_filtering_milliseconds = 2000,
  standard_deviations = 2,
  running_window_width_percentage = 0.025
)
}
\arguments{
\item{data}{Dataframe including participant data, and their ID}

\item{format}{How are the data shaped? "Long" or "Wide"}

\item{unit}{Unit of the response times, can take either "Seconds" or "Milliseconds"}

\item{type_of_filtering}{What type of filtering do you want to apply?  "Fixed_Window" only applies a fixed window type, instead, "Rolling_Standard_Deviation", first applies a fixed windown and then a rolling window}

\item{lower_boundary_fixed_filtering_milliseconds}{During the 'Fixed_Window' filtering what is the lower boundary (in milliseconds)? Enter an integer}

\item{upper_boundary_fixed_filtering_milliseconds}{During the 'Fixed_Window' filtering what is the upper boundary (in milliseconds)? Enter an integer}

\item{standard_deviations}{During the 'Rolling_Standard_Deviation' filtering, how many standard deviations from the mean do you want to filter? Enter a numeric value > 1}

\item{running_window_width_percentage}{During the 'Rolling_Standard_Deviation' filtering, how large is the rolling window? Enter a numeric value from 0 to 1}
}
\value{
A file in wide format with all the reaction time data filtered, and one row for each participant
}
\description{
This function is a wrapper for 'SRTT_analysis' to handle dataframes as input data
The functions goes over the 'data' file.
We expect 'data' to have one column for each subject, with ID on top.
We handle also the option where the format is wide
The first row has subject's ID ("ID")
All the other rows are reaction_times
Units are preferred in milliseconds but we also handle when presented in seconds
The fixed boundary are to remove predictive reaction times (<200ms) or when subject was distracted (>2000ms)
The default value is 2 standard deviation and a window size of ~2.5\% of the total number of trials
}
\examples{
# Basic usage
ID <- LETTERS[1:5]
dataframe_wide <- data.frame(ID = ID, matrix(stats::rnorm(5 * 500, mean = 1, sd = 0.1), nrow = 5, ncol = 500) )
head(SRTT_analysis_for_Dataframes(dataframe_wide, format = "Wide")[, 1:10])

dataframe_long <- as.data.frame(t(dataframe_wide))
head(SRTT_analysis_for_Dataframes(dataframe_long, format = "Long" ,running_window_width_percentage = 0.3)[, 1:10])
}
